feat(T9985): packages/cleo is dispatch-only (E8-CLI-LAYERING)
Move SDK logic from cleo → core per Package-Boundary Check.
Verify dispatch/domains/worktree.ts is router-only (destroyWorktree
re-exported through core, dropping @cleocode/worktree direct dep).
lint-cli-package-boundary baseline tightened (28 → 25).

Migrations:
- backup-inspect tar/manifest helpers → core/store/backup-inspect.ts
- restore-conflicts.md parser → core/store/restore-conflict-parser.ts
- setup --config-json merger → core/setup/config-json-merge.ts

All new SDK primitives are re-exported through @cleocode/core public
barrel so CLI commands import them without violating lint-core-first
RULE-3. Core-first baseline unchanged at 87.

Closes T10040-T10044.
Saga: T9977
Decision: D010
