+++ packages/cleo/src/cli/commands/restore.ts
@@ -17,215 +17,25 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { ExitCode } from '@cleocode/contracts';
-import { CleoError, getProjectRoot, getTaskAccessor } from '@cleocode/core';
+import {
+  CleoError,
+  getProjectRoot,
+  getTaskAccessor,
+  type ParsedResolution,
+  parseConflictReport,
+  setAtPath,
+} from '@cleocode/core';
 import { defineCommand, showUsage } from 'citty';
 import { dispatchRaw } from '../../dispatch/adapters/cli.js';
-import {
-  CLEO_DIR_NAME,
-  RESTORE_CONFLICTS_MD,
-  RESTORE_DEFAULT_FILE,
-  RESTORE_VALID_JSON_FILENAMES,
-} from '../paths.js';
+import { CLEO_DIR_NAME, RESTORE_CONFLICTS_MD, RESTORE_DEFAULT_FILE } from '../paths.js';
 import { cliError, cliOutput, humanLine } from '../renderers/index.js';
 
 // ---------------------------------------------------------------------------
-// Types for conflict report parsing
+// Re-exports for tests + downstream code (T9985 — parser lives in core)
 // ---------------------------------------------------------------------------
 
-/** A single field entry parsed from the conflict report. */
-interface ParsedResolution {
-  /** Which section of the report this came from. */
-  section: 'auto' | 'manual';
-  /** The target JSON file on disk. */
-  filename: 'config.json' | 'project-info.json' | 'project-context.json'; // values from RESTORE_VALID_JSON_FILENAMES
-  /** Dot-separated field path (e.g. "hooks.preCommit"). */
-  fieldPath: string;
-  /** The local (A) value, may be undefined if not present. */
-  localValue: unknown;
-  /** The imported (B) value, may be undefined if not present. */
-  importedValue: unknown;
-  /** The chosen resolution. */
-  resolution: 'A' | 'B' | 'manual-review';
-}
-
-// ---------------------------------------------------------------------------
-// Parser helpers (private to this module)
-// ---------------------------------------------------------------------------
-
-/**
- * Parse a backtick-quoted value from a markdown line such as:
- *   `"openai"` → "openai"
- *   `true`     → true
- *   `42`       → 42
- *   _(not present)_ → undefined
- */
-function parseMarkdownValue(raw: string): unknown {
-  const trimmed = raw.trim();
-  if (trimmed === '_(not present)_' || trimmed === '') return undefined;
-  // Strip surrounding backticks if present
-  const stripped = trimmed.replace(/^`([\s\S]*)`$/, '$1').trim();
-  // Try JSON parse for booleans, numbers, quoted strings, objects, arrays
-  try {
-    return JSON.parse(stripped);
-  } catch {
-    // Return as-is string if JSON parse fails
-    return stripped;
-  }
-}
-
-/**
- * Set a value at a dot-separated path within a plain object tree,
- * creating intermediate objects as needed.
- */
-function setAtPath(obj: Record<string, unknown>, dotPath: string, value: unknown): void {
-  const parts = dotPath.split('.');
-  let curr: Record<string, unknown> = obj;
-  for (let i = 0; i < parts.length - 1; i++) {
-    const key = parts[i] as string;
-    if (curr[key] === undefined || typeof curr[key] !== 'object' || curr[key] === null) {
-      curr[key] = {};
-    }
-    curr = curr[key] as Record<string, unknown>;
-  }
-  const lastKey = parts[parts.length - 1] as string;
-  curr[lastKey] = value;
-}
-
-/**
- * Parse a restore-conflicts.md markdown report into an array of
- * {@link ParsedResolution} entries.
- *
- * The format produced by T357 is:
- * ```
- * ## config.json
- * ### Resolved (auto-applied)
- * - `field.path`
- *   - Local (A): `value`
- *   - Imported (B): `value`
- *   - Resolution: **A**
- *   - Rationale: ...
- * ### Manual review needed
- * - `field.path`
- *   - Local (A): `value`
- *   - Imported (B): `value`
- *   - Resolution: **manual-review**
- *   - Rationale: ...
- * ```
- *
- * @task T365
- * @epic T311
- */
-export function parseConflictReport(md: string): ParsedResolution[] {
-  const results: ParsedResolution[] = [];
-
-  const VALID_FILENAMES = RESTORE_VALID_JSON_FILENAMES;
-
-  // Split into lines for state-machine parsing
-  const lines = md.split('\n');
-
-  let currentFilename: ParsedResolution['filename'] | null = null;
-  let currentSection: 'auto' | 'manual' | null = null;
-
-  // State for the current field entry being accumulated
-  let entryField: string | null = null;
-  let entryLocalRaw: string | null = null;
-  let entryImportedRaw: string | null = null;
-  let entryResolution: 'A' | 'B' | 'manual-review' | null = null;
-
-  /** Flush the current accumulated entry if complete. */
-  function flushEntry(): void {
-    if (
-      entryField !== null &&
-      entryResolution !== null &&
-      currentFilename !== null &&
-      currentSection !== null
-    ) {
-      results.push({
-        section: currentSection,
-        filename: currentFilename,
-        fieldPath: entryField,
-        localValue: entryLocalRaw !== null ? parseMarkdownValue(entryLocalRaw) : undefined,
-        importedValue: entryImportedRaw !== null ? parseMarkdownValue(entryImportedRaw) : undefined,
-        resolution: entryResolution,
-      });
-    }
-    entryField = null;
-    entryLocalRaw = null;
-    entryImportedRaw = null;
-    entryResolution = null;
-  }
-
-  for (const line of lines) {
-    // ## <filename> heading
-    const fileHeading = /^##\s+(.+\.json)\s*$/.exec(line);
-    if (fileHeading) {
-      flushEntry();
-      const name = fileHeading[1]?.trim() ?? '';
-      currentFilename = VALID_FILENAMES.has(name) ? (name as ParsedResolution['filename']) : null;
-      currentSection = null;
-      continue;
-    }
-
-    // ### section heading
-    const sectionHeading = /^###\s+(.+)$/.exec(line);
-    if (sectionHeading) {
-      flushEntry();
-      const headingText = (sectionHeading[1] ?? '').toLowerCase();
-      if (headingText.includes('manual')) {
-        currentSection = 'manual';
-      } else if (headingText.includes('resolved') || headingText.includes('auto')) {
-        currentSection = 'auto';
-      } else {
-        currentSection = null;
-      }
-      continue;
-    }
-
-    if (currentFilename === null || currentSection === null) continue;
-
-    // - `field.path`  — starts a new field entry
-    const fieldLine = /^-\s+`([^`]+)`\s*$/.exec(line);
-    if (fieldLine) {
-      flushEntry();
-      entryField = fieldLine[1] ?? null;
-      continue;
-    }
-
-    if (entryField === null) continue;
-
-    // Sub-bullets:  - Local (A): `value`
-    const localLine = /^\s+-\s+Local\s+\(A\):\s+(.+)$/.exec(line);
-    if (localLine) {
-      entryLocalRaw = localLine[1] ?? '';
-      continue;
-    }
-
-    // - Imported (B): `value`
-    const importedLine = /^\s+-\s+Imported\s+\(B\):\s+(.+)$/.exec(line);
-    if (importedLine) {
-      entryImportedRaw = importedLine[1] ?? '';
-      continue;
-    }
-
-    // - Resolution: **X**
-    const resolutionLine = /^\s+-\s+Resolution:\s+\*\*([^*]+)\*\*/.exec(line);
-    if (resolutionLine) {
-      const resText = (resolutionLine[1] ?? '').trim();
-      if (resText === 'A') {
-        entryResolution = 'A';
-      } else if (resText === 'B') {
-        entryResolution = 'B';
-      } else {
-        entryResolution = 'manual-review';
-      }
-    }
-  }
-
-  // Flush the last accumulated entry
-  flushEntry();
-
-  return results;
-}
+export type { ParsedResolution } from '@cleocode/core';
+export { parseConflictReport } from '@cleocode/core';
 
 // ---------------------------------------------------------------------------
 // Subcommands
